move method to base
